SENTINEL-Backend/requirements.txt
fastapi
uvicorn[standard]
supabase
google-genai
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
python-dotenv
httpx
python-multipart
slack-sdk
